+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Changelog
 
+## Unreleased
+
+### Fixed
+- Fixed a bug where it was possible to add more than one child blocks UI element to a block type's field layout
+
 ## 3.0.0-beta.1 - 2022-04-14
 
 ### Added